% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebird_notsensitive.R
\name{build_predictor_list}
\alias{build_predictor_list}
\title{Removes any predictor names from \code{params$PREDICTOR_LIST} for which the values
are constant in \code{dat}}
\usage{
build_predictor_list(params, dat)
}
\description{
Removes any predictor names from \code{params$PREDICTOR_LIST} for which the values
are constant in \code{dat}
}
