% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_io.R
\name{dat2sf}
\alias{dat2sf}
\title{Convert data frame to sf object with ESPG 4326 SRD}
\usage{
dat2sf(dat, coord_cols = c("longitude", "latitude"))
}
\description{
Convert data frame to sf object with ESPG 4326 SRD
}
