% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_utils.R
\name{data_ecdf}
\alias{data_ecdf}
\title{For each value in one data set, compute its rarity in another}
\usage{
data_ecdf(x, dist)
}
\arguments{
\item{x}{data.frame of interest}

\item{dist}{data.frame from which to compute frequencies}
}
\description{
For each value in one data set, compute its rarity in another
}
