% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{date_ecdf}
\alias{date_ecdf}
\title{Convert a vector of dates to values on \link{0,1}}
\usage{
date_ecdf(dates)
}
\arguments{
\item{dates;}{Date}
}
\description{
Convert a vector of dates to values on \link{0,1}
}
