% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebird_notsensitive.R
\name{detect_species_names}
\alias{detect_species_names}
\title{Given a data frame, return column names that look like species codes}
\usage{
detect_species_names(df)
}
\description{
Given a data frame, return column names that look like species codes
}
