% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_utils.R
\name{get_constant_vars}
\alias{get_constant_vars}
\title{get_constant_vars
Return variable names of columns that are constant across \code{dat}}
\usage{
get_constant_vars(dat)
}
\description{
get_constant_vars
Return variable names of columns that are constant across \code{dat}
}
