% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{get_date_quantiles}
\alias{get_date_quantiles}
\title{get quantiles on vector of dates}
\usage{
get_date_quantiles(dates, probs)
}
\arguments{
\item{dates}{Date vector}

\item{probs}{numeric; quantiles for which you want exceedance probabilities}
}
\description{
get quantiles on vector of dates
}
