% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranger_utils.R
\name{get_ntrees_oob}
\alias{get_ntrees_oob}
\title{Get the number of out-of-bag trees for each data point}
\usage{
get_ntrees_oob(rf)
}
\arguments{
\item{rf}{ranger random forest object (with inbag.counts)}
}
\description{
Get the number of out-of-bag trees for each data point
}
