% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebird_notsensitive.R
\name{get_taxa}
\alias{get_taxa}
\title{Get family and order of birds using the Open Tree of Life
(assumed to be installed and data downloaded to local disk)}
\usage{
get_taxa(sp_codes)
}
\description{
Get family and order of birds using the Open Tree of Life
(assumed to be installed and data downloaded to local disk)
}
