% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebird_notsensitive.R
\name{ith_erd2params}
\alias{ith_erd2params}
\title{Create prepped args for running ebirdstwf::fit_model_stixel() using only a test data frame.}
\usage{
ith_erd2params(
  erd,
  sp,
  stixel_id,
  cci_occ = c(TRUE, FALSE),
  cci_count = c(TRUE, FALSE),
  additional_predictors = NULL,
  config_path = "~/stem_hwf/data_prep/data-prep_configuration.json"
)
}
\arguments{
\item{erd}{data frame with one observation per checklist (wide form, with species as columns)}

\item{sp}{species code to prepare}

\item{stixel_id}{stixel to run}

\item{cci_occ}{Will the predictors be used to run a cci occurrence model?}

\item{cci_count}{Will the predictors be used to run a cci count model?}

\item{additional_predictors}{Character vector of names to add to
PREDICTOR_LIST. Useful when testing a new feature (e.g. "mu_os" for the count
model only).}

\item{config_path}{Path to stem_hwf data-prep_configuration.json.}
}
\description{
Create prepped args for running ebirdstwf::fit_model_stixel() using only a test data frame.
}
