% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebird_notsensitive.R
\name{map_habitats}
\alias{map_habitats}
\title{Map habitats}
\usage{
map_habitats(
  path_erd,
  n_clusters = 8,
  lat_min = 44.13,
  lat_max = 44.46,
  lon_min = -74.22,
  lon_max = -73.33
)
}
\description{
Cluster habitats using kmeans, map them onto a colored plot and describe clusters
}
