% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas.R
\name{pare_formula}
\alias{pare_formula}
\title{Remove terms from list that are constant in the data}
\usage{
pare_formula(dat, predictors)
}
\arguments{
\item{dat}{data frame to be modeled}

\item{predictors}{character vector of predictor names}
}
\description{
Remove terms from list that are constant in the data
}
