% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_pd}
\alias{plot_pd}
\title{plot_pd}
\usage{
plot_pd(pd, include_iqr = TRUE)
}
\arguments{
\item{pd}{data frame with columns "predictor", "x", "summarized_response", "q25_response"
Returns a partial dependence plot given output from \code{calculate_pd()}}
}
\description{
plot_pd
}
