% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranger_utils.R
\name{preds_from_weighted_mean_probs}
\alias{preds_from_weighted_mean_probs}
\title{Get predicted classifications from weighted mean of class assignment probability
(instead of maximum probability across the classes)}
\usage{
preds_from_weighted_mean_probs(pred_mat)
}
\arguments{
\item{pred_mat}{matrix of classification probabilities (n_obs x n_classes)}
}
\description{
Get predicted classifications from weighted mean of class assignment probability
(instead of maximum probability across the classes)
}
