% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_io.R
\name{read_db_table}
\alias{read_db_table}
\title{Get a data frame from a .db file}
\usage{
read_db_table(path_db)
}
\arguments{
\item{dir_db}{path to .db file}
}
\value{
data frame or list of data.frames
}
\description{
Get a data frame from a .db file
}
