% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebird_notsensitive.R
\name{srds2raster}
\alias{srds2raster}
\title{Combine up to multiple SRD predictions objects to a single raster that can be plotted}
\usage{
srds2raster(list_srd_predictions, srd_day, mask_by_binary = TRUE)
}
\description{
Combine up to multiple SRD predictions objects to a single raster that can be plotted
}
