% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_utils.R
\name{stratify2weights}
\alias{stratify2weights}
\title{Create weights inversely proportional to relative class size}
\usage{
stratify2weights(class_col)
}
\arguments{
\item{class_col}{character vector of classes}
}
\description{
Create weights inversely proportional to relative class size
}
