% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranger_utils.R
\name{tree_depth}
\alias{tree_depth}
\title{Get tree depth for each tree
a vector of length num.trees that represents the number of nodes in the longest continuous branch, inclusive of the first and last
#' @export}
\usage{
tree_depth(rf)
}
\description{
Get tree depth for each tree
a vector of length num.trees that represents the number of nodes in the longest continuous branch, inclusive of the first and last
#' @export
}
