% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebird_notsensitive.R
\name{union_stixel_params}
\alias{union_stixel_params}
\title{Extract \verb{params[[@param_name]]} from multiple interactive stixels, and union them into a complete set}
\usage{
union_stixel_params(dir_interactive_stx, stxl_ids = NULL, param_name)
}
\description{
Extract \verb{params[[@param_name]]} from multiple interactive stixels, and union them into a complete set
}
