% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_metrics.R
\name{weighted_mean_se}
\alias{weighted_mean_se}
\title{Calculate a weighted mean and a standard error of the mean in multiple ways}
\usage{
weighted_mean_se(x, obs_ss, type = "2", cap_ss_value = NULL)
}
\description{
Calculate a weighted mean and a standard error of the mean in multiple ways
}
